+++ tagify_perf_test.go
@@ -1,6 +1,7 @@
 package tagify
 
 import (
+	"context"
 	"fmt"
 	"testing"
 )
@@ -11,7 +12,7 @@ func BenchmarkTagify(b *testing.B) {
 	b.ResetTimer()
 
 	for i := 0; i < b.N; i++ {
-		_, err := GetTags(fmt.Sprintf("http://localhost:%d", port), HTML, 5, false, true)
+		_, err := GetTags(context.TODO(), fmt.Sprintf("http://localhost:%d", port), HTML, 5, false, true)
 		if err != nil {
 			break
 		}